added sample jax-rs webapp
