--- samples/jaxrs_basic/src/main/java/demo/jaxrs/client/add_customer.xml
@@ -1,3 +1,0 @@
-<Customer>
-  <name>Jack</name>
-</Customer>
