--- samples/jaxrs_basic/src/main/java/demo/jaxrs/client/update_customer.xml
@@ -1,4 +1,0 @@
-<Customer>
-  <name>Mary</name>
-  <id>123</id>
-</Customer>
