--- samples/jaxrs_basic/src/main/java/demo/jaxrs/server/Server.java
@@ -1,42 +1,0 @@
-/*
- * Copyright 2011-2012 WSO2, Inc. (http://wso2.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package demo.jaxrs.server;
-
-import org.apache.cxf.jaxrs.JAXRSServerFactoryBean;
-import org.apache.cxf.jaxrs.lifecycle.SingletonResourceProvider;
-
-public class Server {
-
-    protected Server() throws Exception {
-        JAXRSServerFactoryBean sf = new JAXRSServerFactoryBean();
-        sf.setResourceClasses(CustomerService.class);
-        sf.setResourceProvider(CustomerService.class, 
-            new SingletonResourceProvider(new CustomerService()));
-        sf.setAddress("http://localhost:9000/");
-
-        sf.create();
-    }
-
-    public static void main(String args[]) throws Exception {
-        new Server();
-        System.out.println("Server ready...");
-
-        Thread.sleep(5 * 60 * 1000);
-        System.out.println("Server exiting");
-        System.exit(0);
-    }
-}
