+++ samples/jaxrs_basic/src/main/java/org/wso2/appserver/sample/service/Order.java
@@ -1,28 +1,36 @@
 /*
- * Copyright 2011-2012 WSO2, Inc. (http://wso2.com)
+ *  Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ *  WSO2 Inc. licenses this file to you under the Apache License,
+ *  Version 2.0 (the "License"); you may not use this file except
+ *  in compliance with the License.
+ *  You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ *  http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Unless required by applicable law or agreed to in writing,
+ *  software distributed under the License is distributed on an
+ *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ *  KIND, either express or implied. See the License for the
+ *  specific language governing permissions and limitations
+ *  under the License.
+ *
  */
 
-package demo.jaxrs.server;
+package org.wso2.appserver.sample.service;
+
+import java.util.HashMap;
+import java.util.Map;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.xml.bind.annotation.XmlRootElement;
-import java.util.HashMap;
-import java.util.Map;
 
+
+/**
+ * Order class
+ */
 @XmlRootElement(name = "Order")
 public class Order {
     private long id;
@@ -54,8 +62,7 @@ public class Order {
     public Product getProduct(
             @PathParam("productId")
             int productId) {
-        System.out.println("----invoking getProduct with id: " + productId);
-        Product p = products.get(new Long(productId));
+        Product p = products.get(Long.valueOf(productId));
         return p;
     }
