+++ samples/jaxrs_basic/src/main/webapp/WEB-INF/beans.xml
@@ -19,16 +19,24 @@
 <beans xmlns="http://www.springframework.org/schema/beans"
        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
        xmlns:jaxrs="http://cxf.apache.org/jaxrs"
-       xsi:schemaLocation="
-         http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-2.0.xsd
-         http://cxf.apache.org/jaxrs http://cxf.apache.org/schemas/jaxrs.xsd">
+       xmlns:context="http://www.springframework.org/schema/context"
+       xsi:schemaLocation="http://cxf.apache.org/jaxrs
+           http://cxf.apache.org/schemas/jaxrs.xsd
+           http://www.springframework.org/schema/beans
+           http://www.springframework.org/schema/beans/spring-beans.xsd
+           http://www.springframework.org/schema/context
+           http://www.springframework.org/schema/context/spring-context.xsd">
 
-    <jaxrs:server id="customerService" address="/customers">
+    <import resource="classpath:META-INF/cxf/cxf.xml"/>
+    <import resource="classpath:META-INF/cxf/cxf-servlet.xml"/>
+
+    <context:component-scan base-package="org.wso2.appserver.sample.service"/>
+
+    <jaxrs:server id="customerWebService" address="/">
         <jaxrs:serviceBeans>
-            <ref bean="serviceBean"/>
+            <ref bean="customerService"/>
         </jaxrs:serviceBeans>
     </jaxrs:server>
 
-    <bean id="serviceBean" class="demo.jaxrs.server.CustomerService"/>
 </beans>
